mbp: Add the BSP_NAME as the TOPDIR function
When it could not find BSP_NAME from command line, then it
will create use the default parent directroy as the BSP_NAME,
and create the whole MBP tree in this directory.
Or user input the BSP_NAME with "-n", then it will create a
new directory in the directory, then create the whole MBP tree
in the new "BSP_NAME" direcotry.

Signed-off-by: Guojian Zhou <guojian.zhou@windriver.com>